Refactor chemistry test modules
